+++ src/document.ts
@@ -20,8 +20,7 @@ interface IDocumentHasResource {
     data: Resource;
     content: 'resource';
 }
-export /* abstract */ class Document
-    implements IDocumentData, IDocumentHasResources, IDocumentHasIds, IDocumentHasId, IDocumentHasResource {
+export class Document implements IDocumentData, IDocumentHasResources, IDocumentHasIds, IDocumentHasId, IDocumentHasResource {
     public data: any;
     public builded: boolean = false;
     public content: any;