+++ src/jsonapi-config.ts
@@ -3,7 +3,7 @@ export class JsonapiConfig {
     public params_separator?: string = '?';
     public unify_concurrency?: boolean = true;
     public cache_prerequests?: boolean = true;
-    /* eslint-disable id-blacklist */
+    /* eslint-disable */
     public parameters?:
         | {
               page: {
@@ -17,5 +17,5 @@ export class JsonapiConfig {
             size: 'page[size]'
         }
     };
-    /* eslint-enable id-blacklist */
+    /* eslint-enable */
 }