+++ src/services/json-ripper.spec.ts
@@ -2,9 +2,12 @@ import { Resource } from '../resource';
 import { JsonRipper } from './json-ripper';
 import { DocumentCollection } from '../document-collection';
 import { TestFactory } from '../tests/factories/test-factory';
+import { IElement } from '../data-providers/data-provider';
+import { ICacheableDocumentResource } from '../interfaces/data-object';
+import { ICacheableDataCollection } from '../interfaces/data-collection';
 
 describe('JsonRipper for resources', () => {
-    let book = TestFactory.getBook('5');
+    let book: any = TestFactory.getBook('5');
     book.attributes.title = 'Fahrenheit 451';
     book.addRelationship(TestFactory.getAuthor('2'), 'author');
     // @todo maxi: factory dont work?
@@ -15,7 +18,7 @@ describe('JsonRipper for resources', () => {
         let mocked_service_data: { [key: string]: any } = { parseToServer: false };
         spyOn(Resource.prototype, 'getService').and.returnValue(mocked_service_data);
 
-        let obj = JsonRipper.toResourceElements('some.key', book);
+        let obj: Array<IElement> = JsonRipper.toResourceElements('some.key', book);
         expect(obj.length).toBe(1);
         expect(obj[0].key).toBe('some.key');
         expect(obj[0].content.data).toMatchObject({
@@ -38,7 +41,7 @@ describe('JsonRipper for resources', () => {
         let mocked_service_data: { [key: string]: any } = { parseToServer: false };
         spyOn(Resource.prototype, 'getService').and.returnValue(mocked_service_data);
 
-        let obj = JsonRipper.toResourceElements('some.key', book, ['author']);
+        let obj: Array<IElement> = JsonRipper.toResourceElements('some.key', book, ['author']);
         expect(obj.length).toBe(2);
         expect(obj[0].key).toBe('some.key');
         expect(obj[1].content.data).toMatchObject({
@@ -55,9 +58,9 @@ describe('JsonRipper for resources', () => {
         let mocked_service_data: { [key: string]: any } = { parseToServer: false };
         spyOn(Resource.prototype, 'getService').and.returnValue(mocked_service_data);
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         await jsonRipper.saveResource(book);
-        let json = await jsonRipper.getResource(JsonRipper.getResourceKey(book));
+        let json: ICacheableDocumentResource = await jsonRipper.getResource(JsonRipper.getResourceKey(book));
         expect(json.data).toMatchObject({
             attributes: { title: /.+/ },
             id: '5',
@@ -76,9 +79,9 @@ describe('JsonRipper for resources', () => {
         let mocked_service_data: { [key: string]: any } = { parseToServer: false };
         spyOn(Resource.prototype, 'getService').and.returnValue(mocked_service_data);
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         await jsonRipper.saveResource(book);
-        let json = await jsonRipper.getResource(JsonRipper.getResourceKey(book));
+        let json: ICacheableDocumentResource = await jsonRipper.getResource(JsonRipper.getResourceKey(book));
         expect(json.data.cache_last_update).toBeGreaterThanOrEqual(Date.now() - 100);
     });
 
@@ -86,9 +89,9 @@ describe('JsonRipper for resources', () => {
         let mocked_service_data: { [key: string]: any } = { parseToServer: false };
         spyOn(Resource.prototype, 'getService').and.returnValue(mocked_service_data);
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: any = new JsonRipper();
         await jsonRipper.saveResource(book, ['author']);
-        let json = await jsonRipper.getResource(JsonRipper.getResourceKey(book), ['author']);
+        let json: any = await jsonRipper.getResource(JsonRipper.getResourceKey(book), ['author']);
         expect(json.included.length).toEqual(1);
         expect(json.included[0]).toMatchObject({
             id: '2',
@@ -104,10 +107,10 @@ describe('JsonRipper for resources', () => {
         let mocked_service_data: { [key: string]: any } = { parseToServer: false };
         spyOn(Resource.prototype, 'getService').and.returnValue(mocked_service_data);
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: any = new JsonRipper();
         book.relationships.author.data = null;
         await jsonRipper.saveResource(book, ['author']);
-        let json = await jsonRipper.getResource(JsonRipper.getResourceKey(book), ['author']);
+        let json: any = await jsonRipper.getResource(JsonRipper.getResourceKey(book), ['author']);
         expect(json.included.length).toEqual(0);
         expect(json.data.relationships.author.data).toEqual(null);
         // expect(json.included[0]).toMatchObject({
@@ -119,7 +122,7 @@ describe('JsonRipper for resources', () => {
     });
 
     it('Requesting DataProvider not cached resource thrown an error', done => {
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         jsonRipper
             .getResource('extrange_type.id')
             .then()
@@ -130,15 +133,15 @@ describe('JsonRipper for resources', () => {
 });
 
 describe('JsonRipper for collections', () => {
-    let authors = new DocumentCollection();
+    let authors: DocumentCollection = new DocumentCollection();
     // TODO: remove books include in next line when toObject gets fixed (call jsonripper in non provided service)
     authors.data.push(TestFactory.getAuthor('2', ['books']));
-    let author1 = TestFactory.getAuthor('1', ['books']);
+    let author1: any = TestFactory.getAuthor('1', ['books']);
     author1.attributes.name = 'Ray Bradbury';
     authors.data.push(author1);
     author1.relationships.books.data[0].id = '1';
     author1.relationships.books.data[1].id = '2';
-    let book1 = author1.relationships.books.data[0];
+    let book1: any = author1.relationships.books.data[0];
     book1.addRelationship(author1, 'author');
 
     /* Is private now
@@ -184,10 +187,10 @@ describe('JsonRipper for collections', () => {
     it('A ripped collection saved via DataProvider is converted to a Json', async done => {
         spyOn(Resource.prototype, 'getService').and.returnValue({});
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         jsonRipper.saveCollection('some/url', authors);
 
-        let json = await jsonRipper.getCollection('some/url');
+        let json: ICacheableDataCollection = await jsonRipper.getCollection('some/url');
         expect(json.data.length).toEqual(2);
         expect(json.data[1]).toMatchObject({
             attributes: { name: 'Ray Bradbury' },
@@ -209,19 +212,19 @@ describe('JsonRipper for collections', () => {
     it('A ripped collection maintain cache_last_update property', async () => {
         spyOn(Resource.prototype, 'getService').and.returnValue({});
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         jsonRipper.saveCollection('some/url', authors);
-        let json = await jsonRipper.getCollection('some/url');
+        let json: ICacheableDataCollection = await jsonRipper.getCollection('some/url');
         expect(json.cache_last_update).toBeGreaterThanOrEqual(Date.now() - 100);
     });
 
     it('A ripped collection with include saved via DataProvider is converted to a Json', async () => {
         spyOn(Resource.prototype, 'getService').and.returnValue({});
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         jsonRipper.saveCollection('some/url/include', authors, ['books']);
 
-        let json = await jsonRipper.getCollection('some/url/include', ['books']);
+        let json: any = await jsonRipper.getCollection('some/url/include', ['books']);
         expect(json.data.length).toEqual(2);
         expect(json.included.length).toEqual(4); // @TODO: equal to 2 when books include is removed in describe's first getAuthor
 
@@ -241,18 +244,18 @@ describe('JsonRipper for collections', () => {
     it('A ripped collection returns cache_last_update on collection and resources property', async () => {
         spyOn(Resource.prototype, 'getService').and.returnValue({});
 
-        let jsonRipper = new JsonRipper();
+        let jsonRipper: JsonRipper = new JsonRipper();
         jsonRipper.saveCollection('some/url/include', authors, ['books']);
 
-        let json = await jsonRipper.getCollection('some/url/include', ['books']);
+        let json: ICacheableDataCollection = await jsonRipper.getCollection('some/url/include', ['books']);
         expect(json.cache_last_update).toBeGreaterThan(0);
 
         // collection.fill responsability to fill, but ripper need to comunicate last update
         expect(json.data[1].cache_last_update).toBeGreaterThan(0);
     }, 50);
 
-    it('Requesting a DataProvider not cached collection thrown an error', async done => {
-        let jsonRipper = new JsonRipper();
+    it('Requesting a DataProvider not cached collection thrown an error', done => {
+        let jsonRipper: JsonRipper = new JsonRipper();
         jsonRipper
             .getCollection('some/bad/url')
             .then()