+++ src/sources/http.service.ts
@@ -14,7 +14,7 @@ export class Http {
     public constructor(private http: HttpClient, private rsJsonapiConfig: JsonapiConfig) {}
 
     public exec(path: string, method: string, data?: IDocumentResource): Observable<IDocumentData> {
-        let req = {
+        let req: { body: IDocumentResource | null; headers: HttpHeaders } = {
             body: data || null,
             headers: new HttpHeaders({
                 'Content-Type': 'application/vnd.api+json',
@@ -25,7 +25,7 @@ export class Http {
         // NOTE: prevent duplicate GET requests
         if (method === 'get') {
             if (!this.get_requests[path]) {
-                let obs = this.http.request<IDocumentData>(method, this.rsJsonapiConfig.url + path, req).pipe(
+                let obs: Observable<IDocumentData> = this.http.request<IDocumentData>(method, this.rsJsonapiConfig.url + path, req).pipe(
                     tap(() => {
                         delete this.get_requests[path];
                     }),